requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# AWS and Cloud
boto3>=1.28.0
botocore>=1.31.0
python-dotenv>=1.0.0

# Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Graph Processing
networkx>=3.1
scikit-learn>=1.3.0

# Weather API
requests>=2.31.0
aiohttp>=3.8.0

# Data Processing
pyarrow>=12.0.0
fastparquet>=2023.7.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Hyperparameter Tuning
ray[tune]>=2.6.0
optuna>=3.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
